requirements.txt
# Supabase Python client
supabase>=2.0.0

# Word document processing
python-docx>=1.1.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (required by supabase)
httpx>=0.25.0

# PostgreSQL direct connection (for DDL operations)
psycopg2-binary>=2.9.9

# OpenAI API for AI-powered editing
openai>=1.0.0

# Streamlit for Web UI
streamlit>=1.28.0
